refactor: set defaultOptions in shared/constants
