+++ test/unit/generators.test.js
@@ -1,5 +1,5 @@
 import _generateServerInjector from '../../src/server/generateServerInjector'
-import { defaultOptions } from '../utils/constants'
+import { defaultOptions } from '../../src/shared/constants'
 import metaInfoData from '../utils/meta-info-data'
 
 const generateServerInjector = (type, data) => _generateServerInjector(defaultOptions, type, data)
